# Optimize fp8 convert time