## Description

Optimize fp8 convert time.

### 1. Why the change?

| model       | main fp8 convert | this PR fp8 convert  | main q8_0 convert |
|-------------|------------------|----------------------|-------------------|
| llama2-7b   | 48s              | 4.13s  (max 21.32s)  | 25s               |
| llama2-13b  | 149s             | 60.33s               | 89s               |
| chatglm2-6b | 49.2s            | 9.21s   (max 31.37s) | 28.2s              |
| gptj-6b     | 151s             | 24.33s (max 83.37s)  | 43s               |